subsys/cmsis_v*: Fix racy mutex testing, decouple from k_mutex
The cmsis implementations of osMutex was trying to inspect internal
k_mutex state (the owner and lock count) in the process of trying to
acquire the lock.  This is unfixably racy, by definition other
contexts will be trying to do the same on the unsynchronized data.

As far as I can tell, the only purpose was to be able to synthesize
osMutex's specified error behavior, which we can do with the existing
return codes from k_mutex_lock().  Add similar logic to osSemaphore,
which didn't have the race but was likewise abusing access to kernel
internals.

(cherry picked from commit 45ea83af89adcb133ea0f4890c1c7331a265051d)

Original-Signed-off-by: Andy Ross <andyross@google.com>
GitOrigin-RevId: 45ea83af89adcb133ea0f4890c1c7331a265051d
Change-Id: I64876ca0ef4ff1dd31d3dc025da5b8a0ace9638b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3960039
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>