twister: cmake: Add prefiltration based on cmake package helper script
Twister allows filtering based on kconfigs and dts, however the
filtration is a part of the cmake stage, i.e. the stage has to pass
first and then twister checks if required properties are available.
This causes problems, when the full cmake stage is unable to pass.
If so, other filtration methods had to be used, e.g. platform_allow.
The commit modifies the twister workflow:
if a test defines filters based on kconfig/dts first a cmake package
helper script is used to extract dt and/or kconfigs and if the
conditions are fulfilled  it proceeds to a regular cmake stage.
If not, test is skipped.

(cherry picked from commit 85dd4975a491e888417301f0c50e5353a1e2ff71)

Original-Co-authored-by: Daniel DeGrasse <daniel.degrasse@nxp.com>
Original-Signed-off-by: Maciej Perkowski <Maciej.Perkowski@nordicsemi.no>
GitOrigin-RevId: 85dd4975a491e888417301f0c50e5353a1e2ff71
Change-Id: I25cee1e8333eb6b4bb4ce8fc200bb64243f72431
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4306926
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>