doc: build: kconfig: expand on multiple symbols
Convert the warning about multiple symbol definitions to a dedicated
section. This allows the warning to be restricted to the information
that is non-intuitive, with a note for actions that should be taken
when working with multiple definitions.

(cherry picked from commit ee1acd8ae512afa82181261f530a3a932a87dc5c)

Original-Signed-off-by: Jordan Yates <jordan.yates@data61.csiro.au>
GitOrigin-RevId: ee1acd8ae512afa82181261f530a3a932a87dc5c
Change-Id: I67dff172e0b526e1c0db33e3975c3fb6ba5b3639
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4978815
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: ChromeOS Prod (Robot) <chromeos-ci-prod@chromeos-bot.iam.gserviceaccount.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>