cmake: gcc: Fix Cortex-R52 FPU type
The commit f10fa0dea8ecea9cbe05e45f42df8e018735d7a5 mapped the
Cortex-R52 processor type to use the "VFPv3" FPU type, but the
toolchain requires the "FPv5" FPU type and refuses to assemble
floating-point instructions when the "VFPv3" FPU type is specified.

This commit updates the build script to specify the FPU type of
`fpv5-sp-d16` when the processor is configured with a single-precision
FPU, and `neon-fp-armv8` when the processor is configured with a
double-precision + Advanced SIMD-capable FPU.

Note that the `fp-armv8` FPU type is an alias for double-precision FPv5
with 32 double-precision registers (refer to the GCC
`gcc/config/arm/arm-cpus.in` for more details); NEON is always
specified in case of a double-precision configuration because the
Cortex-R52 can only be configured as such.

(cherry picked from commit 9f652ea04a6deba0b33c0b34b861195ae967cd7c)

Original-Signed-off-by: Stephanos Ioannidis <stephanos.ioannidis@nordicsemi.no>
GitOrigin-RevId: 9f652ea04a6deba0b33c0b34b861195ae967cd7c
Change-Id: I28f2470a97886b381533401895530308d4b16f38
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3966798
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>