drivers: i2c: esp32: set timeout to allow clock stretching
The ESP32 series MCUs allow to set a timeout which triggers an error
if the SCL line is unchanged for the specified amount of time.

By default, the ESP-IDF HAL sets the timeout to an arbitrary value of
10 times the bus cycle.

This is not sufficient for chips like the TI bq76952, which pulls the
SCL line low (clock stretching) for several 100 µs.

The timeout should also not be dependent on the chosen bitrate, as it
is defined by the time a chip needs for internal calculation before it
can provide requested data or continue communication.

This commit adds a property to devicetree to allow configuration of
the scl timeout. This value is set via direct register access, as the
ESP-IDF HAL does not provide access to the enable bit and does not
give any information about the maximum size of the timeout (defined
in I2C clock cycles in the register).

Fixes #51351

(cherry picked from commit d5168a8d96607d9ca7a560b0c90d5e32be714ffc)

Original-Signed-off-by: Martin Jäger <martin@libre.solar>
GitOrigin-RevId: d5168a8d96607d9ca7a560b0c90d5e32be714ffc
Change-Id: Ic8900dd2011b5d4e2920ad30fe697077afd7ae25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3993388
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>