tests: clock_control: stm32h7_devices: Test perck configuration
When it needs to access perck clock speed, clock_control driver is using
LL API to read RCC registers and compute frequency.
We're using the exact same method to test the frequency and as a result
we were not able to detect that there was an issue when configuring this
clock.

Add a specific case to this test in order to verify perck domain clock if
perck is used in SPI clk configuration.

We're now able to detect issues (and test is failing).

(cherry picked from commit cf84395f7bae67b3aced3607988f91d02e80224e)

Original-Signed-off-by: Erwan Gouriou <erwan.gouriou@linaro.org>
GitOrigin-RevId: cf84395f7bae67b3aced3607988f91d02e80224e
Change-Id: If20da3e70ce3b9628faaff602562cc028157bf3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4020479
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>