net: sockets: socketpair: do not allow blocking IO in ISR context
Using a socketpair for communication in an ISR is not a great
solution, but the implementation should be robust in that case
as well.

It is not acceptible to block in ISR context, so robustness here
means to return -1 to indicate an error, setting errno to `EAGAIN`
(which is synonymous with `EWOULDBLOCK`).

Fixes #25417

(cherry picked from commit d832b04e9627b7ce22a08b610e6828cd73b0d6f7)

Original-Signed-off-by: Chris Friedt <cfriedt@meta.com>
GitOrigin-RevId: d832b04e9627b7ce22a08b610e6828cd73b0d6f7
Change-Id: I5c994710df1d4e04145503bf3462148a4f22ed18
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3979566
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>