i3c: rename is_primary to is_secondary
Rename is_primary to is_secondary. The justification for this is
because it is less likely to have something configured to be
secondary, and the 0 value would be if it is primary.

(cherry picked from commit b7dc01cd48615b9df08fc0e9a0991ee516c18c8a)

Original-Signed-off-by: Ryan McClelland <ryanmcclelland@meta.com>
GitOrigin-RevId: b7dc01cd48615b9df08fc0e9a0991ee516c18c8a
Change-Id: Ia919c8776463d8c3345f650d8386e2f532de9bb5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4101221
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>