tests: Fix coherence of timeout spin lock
Spin locks must be in coherent memory for cavs. Initially this variable
was at the compilation unit scope but warnings about it being unused
from a twister run lead me to move it to be in the ifdef scope in the
function.

Move it back into the compilation units scope and wrap it in an
ifdef to ensure its not labeled as unused.

(cherry picked from commit 6cef9ca5957dd0a145f6fb0f8ad4286f62ac7e53)

Original-Signed-off-by: Tom Burdick <thomas.burdick@intel.com>
GitOrigin-RevId: 6cef9ca5957dd0a145f6fb0f8ad4286f62ac7e53
Change-Id: I71cfbacd8a7c02bbcb835b130227547dafc32a60
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3986355
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: Aaron Massey <aaronmassey@google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>