Bluetooth: Audio: Remove PA sync handling from scan_delegator
The scan delegator will no longer be responsible for
handling the actual synchronization of the periodic
advertisers, and will offload this to higher layers.

The reason for this, is that scanning is a global state,
and should be avoided autonousmly by the stack. The
application is much better suited for handling this.

(cherry picked from commit b37985a898e1ff0f2b147c91dd52b0167f65ff31)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: b37985a898e1ff0f2b147c91dd52b0167f65ff31
Change-Id: I1fd35bab59eccbe4a94e1e22fb63538633fe0b74
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4414165
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Yuval Peress <peress@google.com>