Bluetooth: Host: Fix wrong addr for PA term when using adv list
If using the advertiser list when syncing to a PA, then the
address address needs to be updated for the PA sync object.
This is a small optimization to avoid doing a copy when
the address and SID is already set.

(cherry picked from commit affd8da9a0b0309cd7e23e4550c688e01651a6de)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: affd8da9a0b0309cd7e23e4550c688e01651a6de
Change-Id: I59a7fa11aae3c6ab858b4aed3ca70ad6b1c46c54
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/3990897
Tested-by: Aaron Massey <aaronmassey@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>