arm64: Add comment that `arch_curr_cpu` and `get_cpu` must be synced
The is code duplication as one is in C, and the other is an assembly
macro. As there is no easy way to find out about this duplication,
adding a comment seems the best way to go.

(cherry picked from commit 55df3852bfe151951027f4070e522712dcfa7698)

Original-Signed-off-by: Henri Xavier <datacomos@huawei.com>
GitOrigin-RevId: 55df3852bfe151951027f4070e522712dcfa7698
Change-Id: I98add506099e72f557640849fba2bd64be1c195e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4001682
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>