tests: usb: Make tests for usb mass sample more robust
It was found that the existing regex check was not enough to validate
correctness of execution of mass samples. Tests would pass even when
a filesystem failed to be mounted. An additional line to check is added
to validate the setup of a filesystem.

(cherry picked from commit 15afd318627131ea9dc73ef2f93465bc9aae432a)

Original-Signed-off-by: Maciej Perkowski <Maciej.Perkowski@nordicsemi.no>
GitOrigin-RevId: 15afd318627131ea9dc73ef2f93465bc9aae432a
Change-Id: Ide17d57265ee1de0a9b38de975335482bf602aea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4013212
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>