kernel: Add C++ main() support
The C++ standard requires the main() function to have the return type
of 'int' and does not allow the main() to be defined with the 'void'
return type. Moreover, GCC goes as far as to emit a hard error when the
'::main()' has the return type of `void`.

This commit introduces an option to instruct the Zephyr kernel to call
the 'int main(void)' instead of the 'void main(void)' in case a Zephyr
application defines main() in a C++ source file.

(cherry picked from commit fa5fd41b614cab17da88d2b254a6d44365b8bcd8)

Original-Signed-off-by: Stephanos Ioannidis <stephanos.ioannidis@nordicsemi.no>
GitOrigin-RevId: fa5fd41b614cab17da88d2b254a6d44365b8bcd8
Change-Id: Ia8420737e4f676df6e0462f13e9f88d6cf72261b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4007489
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>