boards: arm: stm32157c-dk2: fix openocd configuration
By default the port 3333 is assigned for the debug of the
Cortex-A CPUx. the port 3334 is assigned for the Cortex-M debug.

This configuration does not allow to debug the Cortex-M.
More than that the stm32mp157 Cortex-A is not supported in zephyr.

This patch introduces openocd configuration, which
allow to use the Zephyr openocd for the Cortex-M.

The configuration consists in disabling the stm32mp15x.cpu<X> port
to set the stm32mp15x.cm4 port to 3333.

(cherry picked from commit ac53c83d42bfa53fe8f70e503adc3750cf9811e8)

Original-Signed-off-by: Arnaud Pouliquen <arnaud.pouliquen@foss.st.com>
GitOrigin-RevId: ac53c83d42bfa53fe8f70e503adc3750cf9811e8
Change-Id: I880f518711731023c261f2714982934e9ebec4f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4030731
Reviewed-by: Yuval Peress <peress@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>