posix: pthread: take care with pthread mutex resources
Previously, `pthread_mutex_init()` could not actually fail, and
destroying mutexes was a no-op, so it was missing in a couple of
places.

However, with the change of `pthread_mutex_t` to `uint32_t`, and
since those are embedded inside of `struct posix_thread` for the
time being, the pthread code needs to keep track that it is
relinquishes used mutex resources when a thread completes.

(cherry picked from commit 6042acc1a93b72a7be01eba7e24902cbf7cbaf72)

Original-Signed-off-by: Chris Friedt <cfriedt@meta.com>
GitOrigin-RevId: 6042acc1a93b72a7be01eba7e24902cbf7cbaf72
Change-Id: I5f0a9403aca0a92f0f056bf9ae25f755f6ac0577
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4028192
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Yuval Peress <peress@google.com>
Tested-by: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>