intel_adsp: ace: power: ipc procedure update
This patch updates ipc response procedure in power down function. New
flow is only limited to the writs into two registers. We need to clear
the IPCxIDD register in case if its contains any leftovers from a
previous responce. And then write a response to the IPCxIDR.

To prepare response we need to copy incoming request and then mark it as
replay. New message with IPC Busy bit set is then send to host.

The reason for this is a change in the behavior of the IPC driver
compared to how it worked when this function was originaly implemented.
The biggest difference are enabled interrupts in register IPCxCTL.

(cherry picked from commit f246d9540c660957b0c41df762382530ffc4c722)

Original-Signed-off-by: Tomasz Leman <tomasz.m.leman@intel.com>
GitOrigin-RevId: f246d9540c660957b0c41df762382530ffc4c722
Change-Id: If6e9755aa6ffc6695caa76f9a43fe74979dc2e22
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4034160
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Reviewed-by: Yuval Peress <peress@google.com>