doc: extensions: kconfig: allow to copy link to results
This patch allows copying a URL to the current Kconfig search results.
For example, I can now share the search results for "CONFIG_TEST.*". A
new button is added to the search box that, when clicked, copies to the
clipboard the URL.

This feature is achieved by prepending query hashes with '!'. When '!'
is found, the content of the URL hash is left intact, whereas before it
was always enclosed within '^$' to produce an exact match.

(cherry picked from commit b38eb53e003bee6afa377dce50a71b75287b2647)

Original-Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
GitOrigin-RevId: b38eb53e003bee6afa377dce50a71b75287b2647
Change-Id: I0433192182310a95d0707ea273b633bc62821726
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4047554
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Keith Short <keithshort@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Yuval Peress <peress@google.com>