tests: clock_control: nrf_onoff_and_bt: Add clearing of `test_end`
This flag needs to be cleared at the beginning of each test. Otherwise,
when the flag remains set after the previous test, the timer handler is
not rescheduled so executing of the current test is actually pointless.

This was in fact the case with `test_bt_interrupted` before switching
the test suite to the new ZTEST API. When this test was executed as
the second one, the timer handler was called only once and the test
could pass (giving false confirmation that the related routines of
the clock_control_nrf driver work properly). After the API switching,
the order of test execution was changed and that test started failing.

(cherry picked from commit a4f063ec491335c1f740cbb9136acbbe8cd12ebf)

Original-Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>
GitOrigin-RevId: a4f063ec491335c1f740cbb9136acbbe8cd12ebf
Change-Id: I68dcb4c523d6effdb4c769a111663b561381a054
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4056938
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>