sys: linear_range: initial API
This patch adds a new API to map values in a linear range to a range
index. This API is useful in the context of regulators, where it is
common that devices map regulator voltages to a set of indices with a
linear relationship.

The original idea is borrowed from Linux.

(cherry picked from commit 01eb1027192391cabee20b8b9c55f82a0c5703bd)

Original-Signed-off-by: Gerard Marull-Paretas <gerard.marull@nordicsemi.no>
GitOrigin-RevId: 01eb1027192391cabee20b8b9c55f82a0c5703bd
Change-Id: I068cc9f64f367d128c214a8869366442b52e9b63
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4055810
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>