include: clock: stm32h5 clock definitions for clock scheme
Defines the clocks for  the stm32H5 device.
The PLL is similar to the stm32U5 except for the mul-n from 8 to 420.
The HSI is similar to the stm32h7 with a prediv.

(cherry picked from commit ff1969de3b6525b2447329fb976302d553aadd20)

Original-Signed-off-by: Francois Ramu <francois.ramu@st.com>
GitOrigin-RevId: ff1969de3b6525b2447329fb976302d553aadd20
Change-Id: Id5e310ba2c6452a46c3d92b28c3cac380f375f42
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4381109
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>