posix: sys/stat.h: adopt the sys/stat.h header from picolibc
This may be getting ahead of the curve a bit, but here we adopt
the `<sys/stat.h>` header from picolibc. It is mostly the same
as that of newlib but the picolibc variant seemed to be slightly
more "inclusive".

Removed the `<_ansi.h>` include.

The commit immediately following this declares a few missing types
only if they are not already declared.

(cherry picked from commit 28726d6f0a03ea0b792ee2948a722bbafed8957c)

Original-Signed-off-by: Chris Friedt <cfriedt@meta.com>
GitOrigin-RevId: 28726d6f0a03ea0b792ee2948a722bbafed8957c
Change-Id: I8e0f7862053c92add47c058bf4358d6542340a9c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4152175
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Tristan Honscheid <honscheid@google.com>