driver: Port80: npcx: defer Port80 code sending to workqueue thread
If the host sends Port80 postcodes frequently while EC is busy handling
other tasks, the Port80 FIFO (16-byte depth) might overflow easily,
especially when the host sends the postcode with the 4-byte format.
This change defers the handling and sending (to the upper layer)
postcodes to the system workqueue thread. It can reduce a lot of
(but not all) the overflow case. Also in practice, we usually care
about the latest postcodes. The older codes are not significant to the
developer. This commit also lowers the printing of the overflow warning
to LOG_DEBUG.

(cherry picked from commit 2e961102c19e79e8739066a4309a95d765e69253)

Original-Signed-off-by: Jun Lin <CHLin56@nuvoton.com>
GitOrigin-RevId: 2e961102c19e79e8739066a4309a95d765e69253
Change-Id: I5706345fe1479a9119eca3a7a5a34818379053f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4271467
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>