nrf52_bsim: Add option to delay CPU and test initialization
In some simulated test cases, it can be usefull
to delay the boot of the CPU and the test initialization
for a while. Namely by the offset of the device
relative to other simulated devices.

We add the option to select so from command line.
By default the option is disabled so this change is
backwards compatible.

(cherry picked from commit fb95defce3a944e2bdfb0d1d5ce20585195c88a5)

Original-Signed-off-by: Alberto Escolar Piedras <alpi@oticon.com>
GitOrigin-RevId: fb95defce3a944e2bdfb0d1d5ce20585195c88a5
Change-Id: I0e117aec6d9f66aa9c3e9ea46c675df3d9a74d1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4111122
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>