Bluetooth: Audio: Fix audio_iso handling for unicast client
For the unicast client, the direction for the endpoint
is reversed in terms of RX/TX, i.e. a sink endpoint is
RX for the unicast server and broadcast sink, but TX
for the unicast client, and similar for the source endpoint.

(cherry picked from commit f239df0ca7bd4f7f66e6756504b72ab801653b04)

Original-Signed-off-by: Emil Gydesen <emil.gydesen@nordicsemi.no>
GitOrigin-RevId: f239df0ca7bd4f7f66e6756504b72ab801653b04
Change-Id: I12bb390397102bd71ae6bbad9e2cad30553a72b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4121378
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>