doc: application development: fix DT overlay section
Commit f17630ba7501cabb40a4d57a4892a1a5bd0603ff
("doc: application: added description of WARN_EXPERIMENTAL setting")
introduced a new section in the Kconfig settings bit about
experimental features.

This broke the structure of the document, however, because it used the
wrong kind of section underline, resulting in a hierarchy that looks
like this:

   Application Configuration
   └── Kconfig Configuration
   Experimental Features
   └── Devicetree Overlays

when it should have looked like this:

   Application Configuration
   ├── Kconfig Configuration
   |   └── Experimental Features
   └── Devicetree Overlays

This falsely make it look like DT overlays are an experimental
feature!

Fix it.

(cherry picked from commit ec725b5cab1933666141403636555948694d4658)

Original-Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>
GitOrigin-RevId: ec725b5cab1933666141403636555948694d4658
Change-Id: I2c4b1e5fa71d5c89bf4878724f705f52f01e8091
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/zephyr/+/4111152
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Al Semjonovs <asemjonovs@google.com>
Reviewed-by: Al Semjonovs <asemjonovs@google.com>
Tested-by: Al Semjonovs <asemjonovs@google.com>